chore: Chore delete old implementation of "Edit Metadata" screen
- Bake in the back confirmation modal in the provider.
