--- mobile/src/navigation/screens/tracks/ModifyViewBase/index.tsx
@@ -1,232 +1,0 @@
-import {
-  MetadataPresets,
-  getMetadata,
-} from "@missingcore/react-native-metadata-retriever";
-import { useNavigation } from "@react-navigation/native";
-import { useEffect } from "react";
-import { useTranslation } from "react-i18next";
-import { BackHandler, View } from "react-native";
-import { KeyboardAwareScrollView } from "react-native-keyboard-controller";
-
-import { Check } from "~/resources/icons/Check";
-import { useFloatingContent } from "../../../hooks/useFloatingContent";
-import type { InitStoreProps, TrackMetadataForm } from "./store";
-import {
-  TrackMetadataStoreProvider,
-  useTrackMetadataIsUnchanged,
-  useTrackMetadataStore,
-} from "./store";
-
-import { ScrollablePresets } from "~/components/Defaults";
-import { Divider } from "~/components/Divider";
-import { ExtendedTButton } from "~/components/Form/Button";
-import { IconButton } from "~/components/Form/Button/Icon";
-import { TextInput } from "~/components/Form/Input";
-import { ModalTemplate } from "~/components/Modal";
-import { StyledText, TEm } from "~/components/Typography/StyledText";
-import { ScreenOptions } from "../../../components/ScreenOptions";
-
-export function ModifyTrackBase(props: InitStoreProps) {
-  const { offset, ...rest } = useFloatingContent();
-  return (
-    <TrackMetadataStoreProvider {...props}>
-      <ScreenConfig />
-      <MetadataForm bottomOffset={offset} />
-      <ConfirmationModal />
-      <ResetWorkflow {...rest} />
-    </TrackMetadataStoreProvider>
-  );
-}
-
-//#region Screen Header
-/** Configuration for the top app bar. */
-function ScreenConfig() {
-  const { t } = useTranslation();
-
-  const isUnchanged = useTrackMetadataIsUnchanged();
-  const trackName = useTrackMetadataStore((s) => s.name);
-  const isSubmitting = useTrackMetadataStore((s) => s.isSubmitting);
-  const onSubmit = useTrackMetadataStore((s) => s.onSubmit);
-
-  return (
-    <ScreenOptions
-      title="feat.trackMetadata.title"
-      // Hacky solution to disable the back button when submitting.
-      headerLeft={isSubmitting ? () => undefined : undefined}
-      headerRight={() => (
-        <IconButton
-          Icon={Check}
-          accessibilityLabel={t("form.apply")}
-          onPress={onSubmit}
-          disabled={
-            isUnchanged || trackName.trim().length === 0 || isSubmitting
-          }
-        />
-      )}
-    />
-  );
-}
-//#endregion
-
-//#region Metadata Form
-function MetadataForm({ bottomOffset }: { bottomOffset: number }) {
-  const { t } = useTranslation();
-
-  const isUnchanged = useTrackMetadataIsUnchanged();
-  const isSubmitting = useTrackMetadataStore((s) => s.isSubmitting);
-  const setShowConfirmation = useTrackMetadataStore(
-    (s) => s.setShowConfirmation,
-  );
-
-  useEffect(() => {
-    const subscription = BackHandler.addEventListener(
-      "hardwareBackPress",
-      () => {
-        if (isUnchanged) return false;
-        if (!isSubmitting) setShowConfirmation(true);
-        return true;
-      },
-    );
-    return () => subscription.remove();
-  }, [isSubmitting, isUnchanged, setShowConfirmation]);
-
-  return (
-    <KeyboardAwareScrollView
-      bottomOffset={16}
-      {...ScrollablePresets}
-      // Remove 24px as `KeyboardAwareScrollView` adds an element at the
-      // end of the ScrollView, causing an additional application of `gap`.
-      contentContainerStyle={{ paddingBottom: bottomOffset - 24 }}
-      contentContainerClassName="gap-6 p-4"
-    >
-      <StyledText dim className="text-center text-sm">
-        {t("feat.trackMetadata.description.line1")}
-        {"\n\n"}
-        {t("feat.trackMetadata.description.line2")}
-      </StyledText>
-      <Divider />
-
-      <View>
-        <TEm textKey="feat.trackMetadata.extra.name" dim />
-        <FormInput field="name" />
-      </View>
-      <View>
-        <TEm textKey="term.artist" dim />
-        <FormInput field="artistName" />
-      </View>
-      <View>
-        <TEm textKey="term.album" dim />
-        <FormInput field="album" />
-      </View>
-      <View>
-        <TEm textKey="feat.trackMetadata.extra.albumArtist" dim />
-        <FormInput field="albumArtist" />
-      </View>
-      <View className="flex-row items-end gap-6">
-        <View className="flex-1">
-          <TEm textKey="feat.trackMetadata.extra.year" dim />
-          <FormInput field="year" numeric />
-        </View>
-        <View className="flex-1">
-          <TEm textKey="feat.trackMetadata.extra.trackNumber" dim />
-          <FormInput field="track" numeric />
-        </View>
-      </View>
-      <View>
-        <TEm textKey="feat.trackMetadata.extra.disc" dim />
-        <FormInput field="disc" numeric />
-      </View>
-    </KeyboardAwareScrollView>
-  );
-}
-
-function FormInput(props: {
-  field: keyof TrackMetadataForm;
-  numeric?: boolean;
-}) {
-  const isSubmitting = useTrackMetadataStore((s) => s.isSubmitting);
-  const field = useTrackMetadataStore((s) => s[props.field]);
-  const setField = useTrackMetadataStore((s) => s.setField);
-
-  return (
-    <TextInput
-      inputMode={props.numeric ? "numeric" : undefined}
-      editable={!isSubmitting}
-      value={field}
-      onChangeText={setField(props.field)}
-      className="w-full border-b border-foreground/60"
-    />
-  );
-}
-//#endregion
-
-//#region Confirmation Modal
-/** Modal that's rendered if we have unsaved changes. */
-function ConfirmationModal() {
-  const navigation = useNavigation();
-  const showConfirmation = useTrackMetadataStore((s) => s.showConfirmation);
-  const setShowConfirmation = useTrackMetadataStore(
-    (s) => s.setShowConfirmation,
-  );
-
-  return (
-    <ModalTemplate
-      visible={showConfirmation}
-      titleKey="form.unsaved"
-      topAction={{
-        textKey: "form.leave",
-        onPress: () => navigation.goBack(),
-      }}
-      bottomAction={{
-        textKey: "form.stay",
-        onPress: () => setShowConfirmation(false),
-      }}
-    />
-  );
-}
-//#endregion
-
-//#region Reset Workflow
-/** Logic to set the form fields to the embedded metadata from the track. */
-function ResetWorkflow({
-  floatingRef,
-  wrapperStyling,
-}: Omit<ReturnType<typeof useFloatingContent>, "offset">) {
-  const { uri } = useTrackMetadataStore((s) => s.initialData);
-  const isSubmitting = useTrackMetadataStore((s) => s.isSubmitting);
-  const setIsSubmitting = useTrackMetadataStore((s) => s.setIsSubmitting);
-  const setFields = useTrackMetadataStore((s) => s.setFields);
-
-  const onReset = async () => {
-    setIsSubmitting(true);
-    try {
-      const trackMetadata = await getMetadata(uri, [
-        ...MetadataPresets.standard,
-        "discNumber",
-      ]);
-      setFields({
-        name: trackMetadata.title ?? "",
-        artistName: trackMetadata.artist ?? "",
-        album: trackMetadata.albumTitle ?? "",
-        albumArtist: trackMetadata.albumArtist ?? "",
-        year: trackMetadata.year?.toString() ?? "",
-        disc: trackMetadata.discNumber?.toString() ?? "",
-        track: trackMetadata.trackNumber?.toString() ?? "",
-      });
-    } catch {}
-    setIsSubmitting(false);
-  };
-
-  return (
-    <View ref={floatingRef} {...wrapperStyling}>
-      <ExtendedTButton
-        textKey="form.reset"
-        onPress={onReset}
-        disabled={isSubmitting}
-        className="bg-red"
-        textClassName="text-neutral100"
-      />
-    </View>
-  );
-}
-//#endregion
